refactor(models): Déplace les modèles veille_marches vers backend pour cohérence API
